MainNavBar component file structure refactoring
